src/data_fetcher.py
class DataFetcher:
    def __init__(self, sources):
        self.sources = sources

    def fetch_data(self):
        all_data = []
        for source in self.sources:
            if source['type'] == 'website':
                data = self._fetch_from_website(source['url'])
            elif source['type'] == 'api':
                data = self._fetch_from_api(source['url'], source.get('params', {}))
            all_data.extend(data)
        return all_data

    def _fetch_from_website(self, url):
        from bs4 import BeautifulSoup
        import requests

        response = requests.get(url)
        if response.status_code == 200:
            soup = BeautifulSoup(response.content, 'html.parser')
            # Implement specific parsing logic based on the website structure
            return self._parse_website_data(soup)
        else:
            return []

    def _fetch_from_api(self, url, params):
        import requests

        response = requests.get(url, params=params)
        if response.status_code == 200:
            return response.json().get('data', [])
        else:
            return []

    def _parse_website_data(self, soup):
        # Placeholder for website-specific parsing logic
        # This should be implemented based on the structure of the website being scraped
        return []