requirements.txt
anyio==4.9.0
APScheduler==3.11.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
python-telegram-bot==22.0
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
soupsieve==2.7
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
